cleaned up and stuff
