+++ Client/Chat_GUI_JFrame.java
@@ -4,21 +4,19 @@
  * and open the template in the editor.
  */
 
-package my.Chat_GUI_JFrame;
 
-import com.sun.glass.events.KeyEvent;
 import java.io.DataInputStream;
 import java.io.DataOutputStream;
 import java.io.IOException;
 import java.io.PipedInputStream;
 import java.io.PipedOutputStream;
-import javax.swing.*;
 import java.awt.event.*;
 
 /**
  *
  * @author Stephen
  */
+@SuppressWarnings("serial")
 public class Chat_GUI_JFrame extends javax.swing.JFrame {
 
     /**
@@ -35,7 +33,6 @@ public class Chat_GUI_JFrame extends javax.swing.JFrame {
      * WARNING: Do NOT modify this code. The content of this method is always
      * regenerated by the Form Editor.
      */
-    @SuppressWarnings("unchecked")
     // <editor-fold defaultstate="collapsed" desc="Generated Code">//GEN-BEGIN:initComponents
     private void initComponents() {
 
@@ -170,16 +167,15 @@ public class Chat_GUI_JFrame extends javax.swing.JFrame {
     }// </editor-fold>//GEN-END:initComponents
 
     private void jTextField1ActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_jTextField1ActionPerformed
-        // TODO add your handling code here:
+        
         
     }//GEN-LAST:event_jTextField1ActionPerformed
 
     private void jTextField2ActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_jTextField2ActionPerformed
-        // TODO add your handling code here:
+       
     }//GEN-LAST:event_jTextField2ActionPerformed
 
     private void jTextField1KeyPressed(java.awt.event.KeyEvent evt) {//GEN-FIRST:event_jTextField1KeyPressed
-        // TODO add your handling code here:
         if(evt.getKeyCode() == KeyEvent.VK_ENTER){
             String line = "";
             line += jTextField2.getText().split(" ")[0] + " ";
