--- SMell06_Server/nbproject/private/private.properties
@@ -1,2 +1,0 @@
-compile.on.save=true
-user.properties.file=C:\\Users\\Stephen\\AppData\\Roaming\\NetBeans\\7.4\\build.properties
