--- SMell06_Server/nbproject/private/private.xml
@@ -1,11 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<project-private xmlns="http://www.netbeans.org/ns/project-private/1">
-    <editor-bookmarks xmlns="http://www.netbeans.org/ns/editor-bookmarks/2" lastBookmarkId="0"/>
-    <open-files xmlns="http://www.netbeans.org/ns/projectui-open-files/2">
-        <group>
-            <file>file:/C:/Users/Stephen/Desktop/Spring_2015/394/ChatProgram/SMell06_Server/src/smell06_server/ClientConnection.java</file>
-            <file>file:/C:/Users/Stephen/Desktop/Spring_2015/394/ChatProgram/SMell06_Server/src/smell06_server/MessageManager.java</file>
-            <file>file:/C:/Users/Stephen/Desktop/Spring_2015/394/ChatProgram/SMell06_Server/src/smell06_server/SMell06_Server.java</file>
-        </group>
-    </open-files>
-</project-private>
